fix: dump sockets before dumping heap
