+++ packages/server/getIsBusy.ts
@@ -1,0 +1,5 @@
+let isBusy = false
+export const getIsBusy = () => isBusy
+export const setIsBusy = (busy: boolean) => {
+  isBusy = busy
+}
