+++ packages/server/graphql/private/mutations/dumpHeap.ts
@@ -2,6 +2,10 @@ import fs from 'fs'
 import inspector from 'inspector'
 import os from 'os'
 import path from 'path'
+import sleep from '../../../../client/utils/sleep'
+import {disconnectAllSockets} from '../../../disconnectAllSockets'
+import {setIsBusy} from '../../../getIsBusy'
+import {Logger} from '../../../utils/Logger'
 import {MutationResolvers} from '../resolverTypes'
 
 const {SERVER_ID} = process.env
@@ -9,6 +13,12 @@ const {SERVER_ID} = process.env
 const dumpHeap: MutationResolvers['dumpHeap'] = async (_source, {isDangerous}) => {
   if (!isDangerous)
     return 'This action will block the server for about 1 minute, Must ack the danger!'
+  Logger.log('[Heap Dump]: Disconnecting sockets')
+  setIsBusy(true)
+  // wait 10 seconds for the readiness probe to fail
+  await sleep(10_000)
+  await disconnectAllSockets()
+  Logger.log('[Heap Dump]: Ready to Dump')
   return new Promise((resolve) => {
     global.gc?.()
     const memoryUsage = process.memoryUsage()
@@ -27,6 +37,7 @@ const dumpHeap: MutationResolvers['dumpHeap'] = async (_source, {isDangerous}) =
     session.post('HeapProfiler.takeHeapSnapshot', undefined, (err) => {
       session.disconnect()
       fs.closeSync(fd)
+      setIsBusy(false)
       resolve(err?.toString() || pathName)
     })
   })
